# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cyber Chaperone Situation Room API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trips
    description: Trip situation room records
  - name: messages
    description: Incoming WhatsApp messages
  - name: webhook
    description: Twilio webhook
  - name: responders
    description: eblockwatch responder network members
  - name: dispatch
    description: Operator dispatch to responders
  - name: case
    description: Situation Room Case participant control

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /webhook/twilio:
    post:
      operationId: twilioWebhook
      tags: [webhook]
      summary: Receive Twilio WhatsApp message
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                Body:
                  type: string
                From:
                  type: string
                To:
                  type: string
                MessageSid:
                  type: string
      responses:
        "200":
          description: TwiML response
          content:
            text/xml:
              schema:
                type: string

  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripBody"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"

  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripBody"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trips/{id}/messages:
    get:
      operationId: getTripMessages
      tags: [trips]
      summary: Get messages for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /trips/{id}/participants:
    get:
      operationId: listCaseParticipants
      tags: [case]
      summary: List Situation Room Case participants for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseParticipant"
    post:
      operationId: inviteCaseParticipant
      tags: [case]
      summary: Invite a participant into the Situation Room Case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteCaseParticipantBody"
      responses:
        "201":
          description: Participant invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseParticipant"

  /trips/{id}/participants/{participantId}:
    patch:
      operationId: updateCaseParticipant
      tags: [case]
      summary: Update participant status (active/removed/declined)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseParticipantBody"
      responses:
        "200":
          description: Updated participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseParticipant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trips/{id}/case-logs:
    get:
      operationId: listCaseLogs
      tags: [case]
      summary: List audit log entries for a Situation Room Case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of case log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseLog"

  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List all incoming messages
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /responders:
    get:
      operationId: listResponders
      tags: [responders]
      summary: List all responders
      responses:
        "200":
          description: List of responders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Responder"
    post:
      operationId: createResponder
      tags: [responders]
      summary: Add a new responder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResponderBody"
      responses:
        "201":
          description: Created responder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responder"

  /responders/{id}:
    patch:
      operationId: updateResponder
      tags: [responders]
      summary: Update a responder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResponderBody"
      responses:
        "200":
          description: Updated responder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Responder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResponder
      tags: [responders]
      summary: Remove a responder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dispatch:
    post:
      operationId: dispatchResponder
      tags: [dispatch]
      summary: Send Situation Room dispatch to a conduit for an active trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchBody"
      responses:
        "200":
          description: Dispatch sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchResponse"
        "404":
          description: Trip or responder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /messages/{id}:
    patch:
      operationId: updateMessage
      tags: [messages]
      summary: Update message (e.g. assign to trip)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageBody"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Trip:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        travelerName:
          type: string
        travelerPhone:
          type: string
        status:
          type: string
          enum: [green, amber, red, completed]
        evidenceNotes:
          type: ["string", "null"]
        inferenceNotes:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        operatorNotes:
          type: ["string", "null"]
        originalMemberEta:
          type: ["string", "null"]
        currentRouteConfidence:
          type: ["string", "null"]
        lastMemberCheckinTime:
          type: ["string", "null"]
          format: date-time
        etaDriftMinutes:
          type: ["integer", "null"]
        iceEscalationStatus:
          type: ["string", "null"]
        startLat:
          type: ["string", "null"]
        startLon:
          type: ["string", "null"]
        destLat:
          type: ["string", "null"]
        destLon:
          type: ["string", "null"]
        routePolyline:
          type: ["string", "null"]
        routeEtaMinutes:
          type: ["integer", "null"]
        routeEtaTime:
          type: ["string", "null"]
        checkpointList:
          type: ["string", "null"]
        messageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - travelerName
        - travelerPhone
        - status
        - evidenceNotes
        - inferenceNotes
        - nextAction
        - operatorNotes
        - messageCount
        - createdAt
        - updatedAt

    CreateTripBody:
      type: object
      properties:
        title:
          type: string
        travelerName:
          type: string
        travelerPhone:
          type: string
        status:
          type: string
          enum: [green, amber, red]
        evidenceNotes:
          type: ["string", "null"]
        inferenceNotes:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        operatorNotes:
          type: ["string", "null"]
      required:
        - title
        - travelerName
        - travelerPhone

    UpdateTripBody:
      type: object
      properties:
        title:
          type: string
        travelerName:
          type: string
        travelerPhone:
          type: string
        status:
          type: string
          enum: [green, amber, red, completed]
        evidenceNotes:
          type: ["string", "null"]
        inferenceNotes:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        operatorNotes:
          type: ["string", "null"]

    Message:
      type: object
      properties:
        id:
          type: integer
        fromNumber:
          type: string
        toNumber:
          type: string
        body:
          type: string
        messageSid:
          type: ["string", "null"]
        tripId:
          type: ["integer", "null"]
        receivedAt:
          type: string
          format: date-time
      required:
        - id
        - fromNumber
        - toNumber
        - body
        - messageSid
        - tripId
        - receivedAt

    UpdateMessageBody:
      type: object
      properties:
        tripId:
          type: ["integer", "null"]

    Responder:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        whatsappNumber:
          type: string
        areaName:
          type: string
        suburb:
          type: ["string", "null"]
        street:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        homeLat:
          type: string
        homeLon:
          type: string
        conduitType:
          type: string
        supportRadiusKm:
          type: ["integer", "null"]
        availabilityStatus:
          type: string
        trustLevel:
          type: string
        linkedNetworkType:
          type: ["string", "null"]
        linkedNetworkName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - whatsappNumber
        - areaName
        - homeLat
        - homeLon
        - conduitType
        - availabilityStatus
        - trustLevel
        - active
        - createdAt
        - updatedAt

    CreateResponderBody:
      type: object
      properties:
        name:
          type: string
        whatsappNumber:
          type: string
        areaName:
          type: string
        suburb:
          type: ["string", "null"]
        street:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        homeLat:
          type: string
        homeLon:
          type: string
        conduitType:
          type: string
        supportRadiusKm:
          type: ["integer", "null"]
        availabilityStatus:
          type: string
        trustLevel:
          type: string
        linkedNetworkType:
          type: ["string", "null"]
        linkedNetworkName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
      required:
        - name
        - whatsappNumber
        - areaName
        - homeLat
        - homeLon

    UpdateResponderBody:
      type: object
      properties:
        name:
          type: string
        whatsappNumber:
          type: string
        areaName:
          type: string
        suburb:
          type: ["string", "null"]
        street:
          type: ["string", "null"]
        province:
          type: ["string", "null"]
        homeLat:
          type: string
        homeLon:
          type: string
        conduitType:
          type: string
        supportRadiusKm:
          type: ["integer", "null"]
        availabilityStatus:
          type: string
        trustLevel:
          type: string
        linkedNetworkType:
          type: ["string", "null"]
        linkedNetworkName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean

    CaseParticipant:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        participantName:
          type: string
        whatsappNumber:
          type: string
        role:
          type: string
        accessStatus:
          type: string
          enum: [invited, active, declined, removed]
        infoLevel:
          type: integer
        permissions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        invitedBy:
          type: ["string", "null"]
        invitedAt:
          type: string
          format: date-time
        removedBy:
          type: ["string", "null"]
        removedAt:
          type: ["string", "null"]
          format: date-time
        responderId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - tripId
        - participantName
        - whatsappNumber
        - role
        - accessStatus
        - infoLevel
        - invitedAt
        - createdAt
        - updatedAt

    InviteCaseParticipantBody:
      type: object
      properties:
        participantName:
          type: string
        whatsappNumber:
          type: string
        role:
          type: string
        infoLevel:
          type: integer
        permissions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        invitedBy:
          type: string
        responderId:
          type: ["integer", "null"]
      required:
        - participantName
        - whatsappNumber
        - role
        - invitedBy

    UpdateCaseParticipantBody:
      type: object
      properties:
        accessStatus:
          type: string
          enum: [invited, active, declined, removed]
        infoLevel:
          type: integer
        permissions:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        removedBy:
          type: ["string", "null"]

    CaseLog:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        participantId:
          type: ["integer", "null"]
        actionType:
          type: string
        operator:
          type: ["string", "null"]
        participantName:
          type: ["string", "null"]
        participantWhatsapp:
          type: ["string", "null"]
        messageSent:
          type: ["string", "null"]
        replyReceived:
          type: ["string", "null"]
        replyCode:
          type: ["string", "null"]
        infoLevelAtTime:
          type: ["integer", "null"]
        tripStatusAtTime:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        operatorNote:
          type: ["string", "null"]
        twilioMessageSid:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tripId
        - actionType
        - createdAt

    DispatchBody:
      type: object
      properties:
        tripId:
          type: integer
        responderId:
          type: integer
        customNote:
          type: ["string", "null"]
        infoLevel:
          type: integer
          description: "1=area only(default), 2=route/dest, 3=member name, 4=member phone"
      required:
        - tripId
        - responderId

    DispatchResponse:
      type: object
      properties:
        sent:
          type: boolean
        messageSid:
          type: ["string", "null"]
        preview:
          type: string
        participantId:
          type: ["integer", "null"]
        caseLogId:
          type: ["integer", "null"]
      required:
        - sent
        - preview