# lib/api-zod/src/generated/types/caseLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyber Chaperone Situation Room API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseLog {
  id: number;
  tripId: number;
  /** @nullable */
  participantId?: number | null;
  actionType: string;
  /** @nullable */
  operator?: string | null;
  /** @nullable */
  participantName?: string | null;
  /** @nullable */
  participantWhatsapp?: string | null;
  /** @nullable */
  messageSent?: string | null;
  /** @nullable */
  replyReceived?: string | null;
  /** @nullable */
  replyCode?: string | null;
  /** @nullable */
  infoLevelAtTime?: number | null;
  /** @nullable */
  tripStatusAtTime?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  operatorNote?: string | null;
  /** @nullable */
  twilioMessageSid?: string | null;
  createdAt: Date;
}