# lib/api-zod/src/generated/types/dispatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cyber Chaperone Situation Room API
 * OpenAPI spec version: 0.1.0
 */

export interface DispatchBody {
  tripId: number;
  responderId: number;
  /** @nullable */
  customNote?: string | null;
  /** 1=area only(default), 2=route/dest, 3=member name, 4=member phone */
  infoLevel?: number;
}